Don't allocate CPU time to idle task when theres work to do

+++ include/arch-ia32/task.h
@@ -3,7 +3,7 @@
 
 #ifndef __ASM__
 struct thread {
-	long esp,eip;
+	uint32_t esp,eip;
 };
 
 /* Actual task switching macro - can't be a function heh */
